backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.7.0
abacusai==1.4.86
sqlalchemy
asyncpg
passlib[bcrypt]
bcrypt<4.1.0
pyjwt
python-multipart
authlib
aiosmtplib
google-auth
httpx
python-dotenv
